+++ lib/src/opengl_controller_native.dart
@@ -1,7 +1,6 @@
 import 'dart:ffi' as ffi;
 import 'dart:io';
 
-import 'flutter_opengl.dart';
 import 'flutter_opengl_ffi.dart';
 import 'opengl_backend.dart';
 import 'opengl_backend_native.dart';
@@ -14,10 +13,9 @@ class OpenGLController {
   OpenGLController._();
 
   late ffi.DynamicLibrary nativeLib;
-  late final FlutterOpengl openglPlugin;
   late final OpenGLBackend openglFFI;
 
-  initializeGL() {
+  void initializeGL() {
     nativeLib = Platform.isAndroid
         ? ffi.DynamicLibrary.open("libflutter_opengl_plugin.so")
         : Platform.isWindows
@@ -25,6 +23,5 @@ class OpenGLController {
             : ffi.DynamicLibrary.process(); // Linux, macOS & iOS
     final ffiObj = FlutterOpenGLFfi.fromLookup(nativeLib.lookup);
     openglFFI = NativeOpenGLBackend(ffiObj);
-    openglPlugin = FlutterOpengl();
   }
 }